backend/requirements.dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements-dev.in -o requirements.dev.txt
asgiref==3.11.1
    # via
    #   django
    #   django-browser-reload
    #   django-cors-headers
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
django==6.0.2
    # via
    #   -r requirements-dev.in
    #   django-admin-sortable2
    #   django-browser-reload
    #   django-cors-headers
    #   django-daisy
    #   django-debug-toolbar
    #   django-storages
    #   django-tinymce
    #   djangorestframework
django-admin-sortable2==2.3.1
    # via -r requirements-dev.in
django-browser-reload==1.21.0
    # via -r requirements-dev.in
django-cors-headers==4.9.0
    # via -r requirements-dev.in
django-daisy==2.0.7
    # via -r requirements-dev.in
django-debug-toolbar==6.2.0
    # via -r requirements-dev.in
django-storages==1.14.6
    # via -r requirements-dev.in
django-tinymce==5.0.0
    # via -r requirements-dev.in
djangorestframework==3.16.1
    # via -r requirements-dev.in
idna==3.11
    # via requests
librt==0.8.0
    # via mypy
mypy==1.19.1
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via mypy
pathspec==1.0.4
    # via mypy
pillow==12.1.1
    # via -r requirements-dev.in
psycopg==3.3.2
    # via -r requirements-dev.in
psycopg-binary==3.3.2
    # via psycopg
requests==2.32.5
    # via -r requirements-dev.in
sqlparse==0.5.5
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.15.0
    # via mypy
urllib3==2.6.3
    # via requests
